Add `AllowPlaceholderSelection` to `TnTInputSelect` class
Modified the `TnTInputSelect` class in the `TnTComponents` namespace. A new parameter `AllowPlaceholderSelection` of type `bool` has been added to control the selection of the placeholder. The condition to check if the `Placeholder` is not null or whitespace has been reformatted for better readability. Depending on the value of `AllowPlaceholderSelection`, the attribute `disabled` or `value` is added to the builder, controlling whether the placeholder can be selected or not.