Update attribute value in TnTInputSelect class
Modified the attribute value in the `TnTInputSelect<TInputType>` class in the `TnTInputSelect.cs` file. The attribute with key 223, which was previously set to `default(TInputType)`, is now set to an empty string. This change could potentially affect the behavior of the program depending on the usage of this attribute in other parts of the code.