Add null check to EditContext in TnTInputBase
This commit adds a null check for `EditContext` in the `TnTInputBase<TInputType>` class within the `TnTInputBase.cs` file. Previously, the `NotifyFieldChanged` method was called on `EditContext` without checking if it was null, which could lead to a `NullReferenceException`. Now, the code ensures `EditContext` is not null before calling the `NotifyFieldChanged` method when the `onblur` event is triggered. This change prevents potential `NullReferenceException` errors.