Update styling for webkit calendar picker in dark theme
Updated the styling of the `::-webkit-calendar-picker-indicator` in a dark theme. The filter to invert the color of the calendar picker indicator is now applied not only when the `#tnt-theme-design-dark` element is present with a media attribute of `(prefers-color-scheme: dark)`, but also when it has a media attribute of `(prefers-color-scheme:dark)`. The filter is now also applied within a `@media (prefers-color-scheme: dark)` block, meaning it will only apply when the user's system is set to a dark color scheme.