Refactor components and classes for isolated JS modules
Significant refactoring was performed on several components and classes to better support isolated JavaScript modules. The `TnTInputCurrency` component was removed from `InputTesting.razor`. The inheritance of several components was changed from `TnTComponentBase` to `TnTPageScriptComponent`. The `TnTPageScript` component was replaced with the `RenderPageScript` method in several `.razor` files. The `JsModulePath` property was changed from `protected` to `public` in several `.razor.cs` files. The `ITnTPageScriptComponent` interface was added, and the `ITnTTextAlign` interface was removed. The `TnTComponentBase` class was significantly refactored, and the `TnTPageScriptComponent` class was added. The `JSRuntimeExt` class was modified to use the new `TnTPageScriptComponent` class. The `RunIsolatedJsScript` property was removed from several classes. The `TnTExternalClickHandler.razor.js` file was modified to add `onLoad`, `onUpdate`, and `onDispose` functions. The `DisposeAsync` and `OnAfterRenderAsync` methods were modified in several classes to use the new `IsolatedJsModule` property. The `TnTLayout.razor.cs`, `TnTSideNavMenuGroup.razor.cs`, and `TnTTabView.razor.cs` files were modified to remove the `RunIsolatedJsScript` property and add a new `JsModulePath` property.