Add Debouncer.cs with debouncing functionality
Added the `Debouncer.cs` file to the `TnTComponents.Core` namespace. This file introduces the `Debouncer` class which is used to debounce asynchronous and synchronous actions, preventing multiple calls of the same action in a short period of time. The class implements the `IDisposable` interface and includes a `Dispose` method to free up resources. It also contains a private field `_debounceCancellationTokenSource` for cancelling debouncing operations. The class provides two `Debounce` methods and two `Cancel` methods. The file includes a MIT License and a link to the original source code on GitHub.