Added JSInterop and resize functionality to TnTAccordion
Added the `Microsoft.JSInterop` namespace to the `TnTAccordion.razor.cs` file to enable .NET code to call JavaScript functions. A new `Resize()` method was added to the `TnTAccordion` class, which invokes the `resize` JavaScript function from the `IsolatedJsModule` and passes the `Element` as a parameter. If the `IsolatedJsModule` is null, it returns a completed `ValueTask`. Also, a new JavaScript function `resize` was added to the `TnTAccordion.razor.js` file. This function takes an `element` and a `dotNetRef` as parameters and calls the `open` method on the `element` if it exists and is open. This function is exported for external use.