Update attribute logic in TnTInputBase.cs
Updated the logic for setting the `readonly` and `disabled` attributes in `TnTInputBase.cs`. Previously, these attributes were set to either their parent form attribute or their own attribute, whichever was not null. Now, if the parent form attribute has a value, the attribute is set to the logical OR of its own value and the parent form attribute's value. If the parent form attribute does not have a value, the attribute is set to its own value.