`Update DisposeAsync methods and adjust method calls`
Refactored the `DisposeAsync` method in `TnTExternalClickHandler.razor.cs` to override the base class method instead of hiding it. Adjusted the call to `IsolatedJsModule.InvokeVoidAsync` in the same file by removing the `Element` parameter. Replaced the call to `IsolatedJsModule.DisposeAsync` with a call to `base.DisposeAsync`. In `TnTPageScriptComponent.cs`, changed the `DisposeAsync` method to `virtual async` to allow overriding.