Updated TnTAccordion class and related JS functions
Updated the `TnTAccordion` class in `TnTAccordion.razor.cs` to call a new JavaScript function `resizeAccordion` instead of `resize`. Added two new methods `Close` and `Open` that call the JavaScript functions `closeAccordion` and `openAccordion` respectively. In `TnTAccordion.razor.js`, renamed the `resize` function to `resizeAccordion` and added `closeAccordion` and `openAccordion` functions. The `closeAccordion` function checks if the element has a `close` method and calls it if it exists. The `openAccordion` function calls the `resizeAccordion` function. In `_accordion.scss`, added a new CSS rule to the `span.tnt-icon` selector to set the `color` property to `inherit`.