Added default item count for virtualized data grid
A new condition has been added to the `ResolveItemsRequestAsync` method in the `TnTDataGrid<TGridItem>` class. This condition checks if `Virtualize` is `true` and `request.Count` is `null`. If these conditions are met, `request.Count` is set to `1`. This change handles scenarios where the data grid is set to virtualize its items but the count of items to be fetched is not specified, defaulting to fetching one item.