Add `ShouldHavePlaceholderSelected` param to `TnTInputSelect`
The `TnTInputSelect` class has been updated to include a new parameter `ShouldHavePlaceholderSelected`, which is set to `true` by default. This parameter determines whether the placeholder should be selected or not. The `selected` attribute of the `option` element is now conditionally added based on the value of `ShouldHavePlaceholderSelected`. Additionally, the `value` attribute of the `option` element has been modified. If the type of `TInputType` is `string`, the `value` attribute is set to an empty string. Otherwise, it is set to the default value of `TInputType`.