Refactored code and updated styles for scheduler components
In this commit, several changes were made across multiple files. In `ScheduleViewBase.cs`, `SchedulerBase.cs`, `TnTEvent.cs`, and `TnTScheduler.razor.cs`, the `System` namespaces were removed, properties and methods were rearranged, and certain methods were moved down. A new method, `CreateEventAsync`, was added to `ScheduleViewBase.cs` to add new events to the scheduler. An `Element` property was added to `SchedulerBase.cs`. In `TnTWeekView.razor`, a `foreach` loop was added to display the events in the scheduler. In `TnTWeekView.razor.cs`, the `Microsoft.AspNetCore.Components` namespace was removed and the `GetVisibleDates` method was moved to the bottom of the class. Lastly, in `_scheduler.scss`, new styles were added to the `.tnt-week-view` and `.tnt-event` classes, and the `&:hover` style was moved down in the `.tnt-day-slot:not(.tnt-header-slot)` class.