`Updated GetVisibleDates() method in TnTWeekView class`
` `

`Modified the GetVisibleDates() method in the TnTWeekView class to improve the calculation of the start date. The start date is now calculated by first determining the difference between StartWeekDay and Scheduler.StartDate.DayOfWeek, and then adding this difference to Scheduler.StartDate. This change is expected to provide a more accurate start date based on the current day of the week.`