Refactor TnTWeekView to use Scheduler object
This commit includes several changes to the `TnTWeekView.razor` and `TnTWeekView.razor.cs` files. The `DisplayedView` and `TasksList` properties are now accessed through a `Scheduler` object instead of being directly accessed. This change is reflected in all conditional checks and loops where these properties are used. The `DisplayedView` and `TasksList` cascading parameters have been removed from the `TnTWeekView.razor.cs` file. The `TasksList` is now sorted by `DateStart` and then by `DateEnd` in descending order each time the view is redrawn. The `isHidden` string is now set based on the `DisplayedView` property of the `Scheduler` object.