Updated task scheduling and UI enhancements
In `Scheduler.razor`, uncommented tasks were added to the `TasksList` and task dates were adjusted. The `EventComparer` class in `Tasks.cs` was removed. `TnTWeekView.razor` saw updates to div elements for improved drag and drop functionality and style changes. The `tnt-label-medium` class was added to a div and the `draggable` attribute was set based on the `DisableDragAndDrop` property. `_scheduler.scss` was updated to prevent user text selection and improve `.tnt-event` class styling.