`Add color customization to TnTScheduler component`
Updated the `TnTScheduler` class to include two new properties, `BackgroundColor` and `TextColor`, both of type `TnTColor`. These properties are decorated with the `[Parameter]` attribute, allowing them to be used as parameters in the component. The `CssClass` property has been updated to include these new properties in the `CssClassBuilder`, enabling the CSS class for the `TnTScheduler` component to include styles for background color and text color. This update allows users to customize the background and text colors of the `TnTScheduler` component.