Update conditional inclusion of `BuildWebCompiler2022`
Updated the `BuildWebCompiler2022` package reference in the `TnTComponents.csproj` file to be conditionally included based on the operating system. This package is now only included when the operating system is Windows NT, as opposed to its previous unconditional inclusion. This change was made to address potential compatibility issues with non-Windows NT operating systems.