Added InfiniteScroll page and updated TnTVirtualize component
A new InfiniteScroll page was added in `InfiniteScroll.razor` using the `TnTVirtualize` component for infinite scrolling functionality. A navigation link for this page was added in `MainLayout.razor`. The `TnTVirtualize` component was updated with support for infinite scrolling and customizable templates. The `TnTVirtualize` class in `TnTVirtualize.razor.cs` was refactored, removing inheritance from `TnTPageScriptComponent` and updating several methods. The `TnTVirtualizeItemsRequestDelegate` in `TnTVirtualizeItemsRequest.cs` was updated to take a `TnTVirtualizeItemsRequest` parameter. The solution configuration platforms in `TnTComponents.sln` were updated to include Debug and Release for Any CPU.