Updated icon rendering and stylesheets integration
Updated the `App.razor` file to include new stylesheets from Google Fonts. Replaced `@Icon` with `@Icon.Render()` in multiple `.razor` files, indicating a new `Render` method in the `Icon` object. Replaced `@MaterialIcon` with a `MaterialIcon` component with an `Icon` property in several `.razor` files. Added a new `Render` method to the `TnTIcon` class in `TnTIcon.cs`. Updated import statements in `_materialIcons.scss` to use the `url` function. Changed the class of the `span` element in `TnTThemeToggle.razor` and updated the `themeSelected` function in `TnTThemeToggle.razor.js`.