`Update theme selection and initialization in TnTThemeToggle`
Updated `TnTThemeToggle.razor.js` to enhance theme selection and initialization. The `themeSelected(e)` function now calls `initSelect(currentTheme)` on the parent element of the target, if it exists. The `onLoad(element, dotNetElementRef)` function has been modified to take `currentTheme` as a parameter in `initSelect()`, and to select the stored theme option within the 'select' element. The `onUpdate(element, dotNetElementRef)` function now calls `updateIcon(currentTheme)` and `initSelect(currentTheme)` on the element, if these methods exist. In `TnTThemeDesign.cs`, the code adding the "media" attribute to the builder has been commented out.