`Add conditional rendering for dates in TnTWeekView`
` `

`Introduced a conditional rendering of the date in TnTWeekView.razor, which is now only displayed if the HideDates property is set to false. Added a new HideDates property as a Parameter in TnTWeekView.razor.cs, allowing the value to be passed from a parent component. This gives the parent control over whether the dates should be displayed or not in the TnTWeekView component.`