`Added DialogHelperComponent and updated TnTDialogService`
Added a new class `DialogHelperComponent` in `DialogHelperComponent.cs` that inherits from `ComponentBase` and includes a `RenderFragment` parameter `ChildContent`. The `BuildRenderTree` method of this class adds the `ChildContent` to the `RenderTreeBuilder`.

In `TnTDialogService.cs`, the `TnTComponents.Dialog.Infrastructure` namespace has been imported. Two new methods, `OpenAsync` and `OpenForResultAsync`, have been added to the `TnTDialogService` class. Both methods accept a `RenderFragment` and `TnTDialogOptions` as parameters, throw an `ArgumentNullException` if the `RenderFragment` is null, and call the corresponding method of the `DialogHelperComponent` class, passing the `RenderFragment` as the `ChildContent`. The `OpenForResultAsync` method returns a `DialogResult`.