Replace `Move` task with `Copy` in `MoveCSS` target
In the `TnTComponents.csproj` file, the `Move` task in the `MoveCSS` target has been replaced with the `Copy` task. This change affects the handling of the `style.css` file during the build process. Previously, the file was moved from the `Styles` directory to the `wwwroot` directory. Now, a copy of the `style.css` file is made in the `wwwroot` directory, leaving the original `style.css` file in the `Styles` directory unchanged after the build.