Update CSS properties in layout and typeahead stylesheets
In `_layout.scss`, the `min-height` property of `tnt-body` has been reduced from `100vh` to `30vh` under three different scenarios. A missing closing brace was also added. In `_typeahead.scss`, the `margin`, `z-index`, and `position` properties of an unnamed element were updated, and a missing closing brace was added. In `style.scss`, a `background-color` property was added to the `body` selector.