Add Description property to TnTEvent and conditional rendering in TnTWeekView
A new nullable property `Description` has been added to the `TnTEvent` record in `TnTEvent.cs`. In `TnTWeekView.razor`, a conditional rendering block has been implemented. If the `Description` property of the `@event` object is not null or whitespace, a paragraph element with the class `tnt-body-small` is rendered, displaying the `Description` property of the `@event` object.