Add DisableToggle param to TnTChip and update styles
Introduce a new `DisableToggle` parameter to the `TnTChip` class, allowing the toggle functionality to be disabled. Update the `ElementClass` property to conditionally include the "tnt-togglable" class based on this parameter. Remove the `BackgroundColor` parameter from the `TnTChip` class. Modify the rendering logic to conditionally add the "checked" attribute and the "onchange" event handler based on the `DisableToggle` parameter. Update `_chips.scss` to apply styles conditionally based on the "tnt-togglable" class when the input is checked.