Enhance input UX and programmatic focus in typeahead
Improved the user experience in `_input.scss` by preventing text
selection within the input and changing the cursor to a text
cursor. Added `@ref` attribute to the input element in
`TnTTypeahead.razor` for programmatic access. Introduced a new
private field `_inputBox` in `TnTTypeahead.razor.cs` and
implemented programmatic focus on the input box after an item
is selected.